refactor SHA256 padding and endianness handling in OpenCL implementation

+++ src/opencl/headers/sha256/sha256.h
@@ -1,1 +1,1 @@
-void sha256(uchar *message, uint messageLength, uchar *hashedMessage);
+void sha256(uchar *message, ulong messageLength, uchar *hashedMessage);
